+++ common/src/main/java/com/phantomwing/thesilverage/item/ModItems.java
@@ -109,6 +109,16 @@ public class ModItems {
     public static final RegistrySupplier<Item> WAXED_WEATHERED_SILVER_BRICK_STAIRS = registerBlock("waxed_weathered_silver_brick_stairs", ModBlocks.WAXED_WEATHERED_SILVER_BRICK_STAIRS);
     public static final RegistrySupplier<Item> WAXED_OXIDIZED_SILVER_BRICK_STAIRS = registerBlock("waxed_oxidized_silver_brick_stairs", ModBlocks.WAXED_OXIDIZED_SILVER_BRICK_STAIRS);
 
+    // Silver Brick Wall
+    public static final RegistrySupplier<Item> SILVER_BRICK_WALL = registerBlock("silver_brick_wall", ModBlocks.SILVER_BRICK_WALL);
+    public static final RegistrySupplier<Item> EXPOSED_SILVER_BRICK_WALL = registerBlock("exposed_silver_brick_wall", ModBlocks.EXPOSED_SILVER_BRICK_WALL);
+    public static final RegistrySupplier<Item> WEATHERED_SILVER_BRICK_WALL = registerBlock("weathered_silver_brick_wall", ModBlocks.WEATHERED_SILVER_BRICK_WALL);
+    public static final RegistrySupplier<Item> OXIDIZED_SILVER_BRICK_WALL = registerBlock("oxidized_silver_brick_wall", ModBlocks.OXIDIZED_SILVER_BRICK_WALL);
+    public static final RegistrySupplier<Item> WAXED_SILVER_BRICK_WALL = registerBlock("waxed_silver_brick_wall", ModBlocks.WAXED_SILVER_BRICK_WALL);
+    public static final RegistrySupplier<Item> WAXED_EXPOSED_SILVER_BRICK_WALL = registerBlock("waxed_exposed_silver_brick_wall", ModBlocks.WAXED_EXPOSED_SILVER_BRICK_WALL);
+    public static final RegistrySupplier<Item> WAXED_WEATHERED_SILVER_BRICK_WALL = registerBlock("waxed_weathered_silver_brick_wall", ModBlocks.WAXED_WEATHERED_SILVER_BRICK_WALL);
+    public static final RegistrySupplier<Item> WAXED_OXIDIZED_SILVER_BRICK_WALL = registerBlock("waxed_oxidized_silver_brick_wall", ModBlocks.WAXED_OXIDIZED_SILVER_BRICK_WALL);
+
     // Cut Silver Slab
     public static final RegistrySupplier<Item> CUT_SILVER_SLAB = registerBlock("cut_silver_slab", ModBlocks.CUT_SILVER_SLAB);
     public static final RegistrySupplier<Item> EXPOSED_CUT_SILVER_SLAB = registerBlock("exposed_cut_silver_slab", ModBlocks.EXPOSED_CUT_SILVER_SLAB);